+++ iblrig/test/test_online_plots.py
@@ -45,3 +45,46 @@ class TestOnlinePlots:
     def test_from_existing_file(self, task_file, qtbot):
         view = op.OnlinePlotsView(task_file)
         assert view.model._n_trials > 0
+
+    def test_colors(self, task_file, qtbot):
+        view = op.OnlinePlotsView(task_file.parent)
+        model = view.model
+        model._trial_data['response_time'] = 1
+        model._seconds_elapsed = 0
+        model._trials_elapsed = 0
+        model.compute_end_session_criteria()
+        assert model.titleColor == op.Colors.TRANSPARENT
+
+        # if the mouse has been training for more than 90 minutes subject training too long -> RED
+        with qtbot.assertNotEmitted(model.titleColorChanged):
+            model._seconds_elapsed = 90 * 60
+            model.compute_end_session_criteria()
+        with qtbot.waitSignal(model.titleColorChanged, timeout=5):
+            model._seconds_elapsed += 1
+            model.compute_end_session_criteria()
+            assert model.titleColor == op.Colors.RED
+
+        # the mouse fails to do more than 400 trials in the first 45 mins -> GREEN
+        model._seconds_elapsed = 45 * 60
+        model.compute_end_session_criteria()
+        with qtbot.assertNotEmitted(model.titleColorChanged):
+            model._n_trials_engaged = 401
+            model.compute_end_session_criteria()
+        with qtbot.waitSignal(model.titleColorChanged, timeout=5):
+            model._n_trials_engaged -= 1
+            model.compute_end_session_criteria()
+            assert model.titleColor == op.Colors.GREEN
+
+        # reaction time over last 20 trials is more than 5 times greater than the overall reaction time -> YELLOW
+        model._n_trials_engaged = 401
+        model.compute_end_session_criteria()
+        with qtbot.assertNotEmitted(model.titleColorChanged):
+            model._trial_data['response_time'][-20:] = 5
+            model.compute_end_session_criteria()
+            assert model.titleColor == op.Colors.TRANSPARENT
+        with qtbot.waitSignal(model.titleColorChanged, timeout=5):
+            model._trial_data['response_time'][-20:] = 5.1
+            model.compute_end_session_criteria()
+            assert model.titleColor == op.Colors.YELLOW
+
+        model.jsonableWatcher.removePath(str(task_file))